Rename StepContext#getJobId to StepContext#getJobInstanceId
There is no such concept of job Id. A job has a name but not an Id.
A job instance however has an Id. The goal of the getter suggested in
BATCH-2769 is to access the JobInstance Id.

This commit renames the newly added StepContext#getJobId to
StepContext#getJobInstanceId which is more explicit.

Resolves BATCH-2769